packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "google-adk>=1.11.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]
